feat: 写完全部函数的代码

+++ modules/nal2/android/src/main/java/com/nal2/Nal2Module.kt
@@ -713,7 +713,7 @@ class Nal2Module(reactContext: ReactApplicationContext) : ReactContextBaseJavaMo
                         data,
                         acDouble,
                         bcDouble,
-                        level,
+                        level.toDouble(),
                         limiting,
                         channels,
                         direction,
